Update edge drawing model
Edge drawing no longer calculated just-in-time using a DrawLine call for
each grid direction satisfying a mask check. Now this calculation is
done ahead of time (at mapgen) and the index of the correct texture is
stored.

Textures exist on a per-vertex basis, rather than a per-face basis,
because this reduces the number of cases that need checking. Mapgen also
updated to disallow clusters of 3 adjacent faces that all have different
heights (to further reduce the number of cases).

This change dramatically reduces the complexity of the codebase and
allows the removal of many ad-hoc functions, improving runtime overall.